+++ pkg/util/json.go
@@ -18,48 +18,43 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 
-package v1
+package util
 
 import (
-	"encoding/base64"
 	"encoding/json"
 
-	"github.com/arangodb/kube-arangodb/pkg/util"
-	"github.com/arangodb/kube-arangodb/pkg/util/errors"
+	"sigs.k8s.io/yaml"
 )
 
-var _ json.Marshaler = &Data{}
-var _ json.Unmarshaler = &Data{}
-
-type Data []byte
-
-func (d Data) MarshalJSON() ([]byte, error) {
-	s := base64.StdEncoding.EncodeToString(d)
-
-	return json.Marshal(s)
-}
-
-func (d Data) SHA256() string {
-	return util.SHA256(d)
-}
-
-func (d *Data) UnmarshalJSON(bytes []byte) error {
-	if d == nil {
-		return errors.Errorf("nil object provided")
-	}
-
-	var s string
-
-	if err := json.Unmarshal(bytes, &s); err != nil {
-		return err
-	}
-
-	ret, err := base64.StdEncoding.DecodeString(s)
+func JSONRemarshal[A, B any](in A) (B, error) {
+	d, err := json.Marshal(in)
 	if err != nil {
-		return err
+		return Default[B](), err
 	}
 
-	*d = ret
+	var o B
 
-	return nil
+	if err := json.Unmarshal(d, &o); err != nil {
+		return Default[B](), err
+	}
+
+	return o, nil
+}
+
+func JsonOrYamlUnmarshal[T any](b []byte) (T, error) {
+	var z T
+
+	if json.Valid(b) {
+		if err := json.Unmarshal(b, &z); err != nil {
+			return Default[T](), err
+		}
+
+		return z, nil
+	}
+
+	if err := yaml.UnmarshalStrict(b, &z); err != nil {
+		return Default[T](), err
+	}
+
+	return z, nil
 }